rename files with unnecessary silly names
Rename files I a postfixed with 'mk' to prevent collisions with standard
python library files.  Turns out, a module was the correct way to fix
this problem instead!
